atmega scale and lora, teensy sleep mode with lora clean wake up

+++ Firmware/ArduinoProMini/main/memorycard.h
@@ -13,7 +13,7 @@ class MemoryCard {
   
 public:
   static void init();
-  static void write_data(const char* fileName, float* xValues, float* yValues, unsigned int size, unsigned long captureTime);
+  static void write_data(const char* fileName, int* xValues, int* yValues, unsigned int size, unsigned int captureTime);
   static void read_data(const char* fileName);
 };
